backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
fastapi-cors
pydantic
groq
python-dotenv
torch
torchvision
tqdm
chromadb
sentence-transformers
requests